requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fireworks-ai==0.14.0
Flask==3.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pillow==10.3.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.1.0
Werkzeug==3.0.1
